+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2025.09.02
+### feat: Use rule's short description
+- Use rule's provided short description, if possible
 ## 2025.09.01
 ### feat: Publish release
 - feat: Publish release